silence warning by reordering switch
On Octave calling fieldnames on a classdef is a warning.  Put that
case last in chebfun/get so that we don't trigger the warning when
accessing more common public properties.

I assume this is not recommended OOP practice so probably this is
a good chance on Matlab too.  Or maybe it should raise a warning
on Matlab?